+++ CHANGELOG.md
@@ -27,6 +27,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Add refs information to commit history views
 - Add 'refresh git history' button (#575) by @nicoco
 - Add settings-alias support (#581) by @nitbook
+- Add 'open current file' button (#583) by @nicoco
 
 ## Removed